split components/Blog directory
