+++ src/components/Blog/post/FigureImage.js
